requirements.txt
av==12.1.0
djitellopy==2.5.0
numpy==1.26.4
opencv-python==4.10.0.82
pillow==10.3.0
# py-nmclient @ file:///./wheels/py_nmclient-0.1.0-py3-none-any.whl
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
shiboken6==6.7.0
